pyproject.toml
[project]
name = "todo-cli-app"
version = "0.1.0"
description = "A command-line todo application with in-memory storage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
todo = "todo_cli.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"